# By Rivet-the-Zombie (14) and others
# Via KA101 (11) and others
* upstream/master: (68 commits)
  Price note for single malt whiskey
  Fixes descriptions
  Fix drawing screwups in construction_menu().
  Don't blank the dialog inside the key reading loop.
  Digestion improvements
  Carnivores can eat dairies!
  Fix note colors on minimap
  fix copy-paste glitch
  Adds a second material type to most comestibles.
  Fixed problems pointed out by KA101
  added how to use to note popup
  Rename veh_part to cargo_part in game::pickup.
  colored map notes
  GLIBCXX_DEBUG: Recompute indices every time in case armor was destroyed in previous loop iterations.
  GLIBCXX_DEBUG: Don't use veh_part for two things - causes read of invalid element.
  GLIBCXX_DEBUG: If we run out of hotkeys for construction, print space instead of reading arbitrary data..
  Aaand now I'm crashing making arrows. Do the same treatment for inventory::use_charges().
  Whoops, actually need to check the precondition ourselves when doing stepping manually.
  Rework fix to address comments by BevapDin.
  GLIBCXX_DEBUG fix: Don't walk off the front of the iterators when crafting using the first item.
  ...

Conflicts:
	data/json/items/armor.json
	src/inventory.cpp

+++ src/overmapbuffer.cpp
@@ -299,7 +299,7 @@ std::vector<point> overmapbuffer::find_all(const tripoint& origin, const std::st
 
 npc* overmapbuffer::find_npc(int id) {
     for (std::list<overmap>::iterator it = overmap_list.begin(); it != overmap_list.end(); ++it) {
-        for (int i = 0; i < it->npcs.size(); i++) {
+        for (size_t i = 0; i < it->npcs.size(); i++) {
             if (it->npcs[i]->getID() == id) {
                 return it->npcs[i];
             }
@@ -311,7 +311,7 @@ npc* overmapbuffer::find_npc(int id) {
 void overmapbuffer::remove_npc(int id)
 {
     for (std::list<overmap>::iterator it = overmap_list.begin(); it != overmap_list.end(); ++it) {
-        for (int i = 0; i < it->npcs.size(); i++) {
+        for (size_t i = 0; i < it->npcs.size(); i++) {
             npc *p = it->npcs[i];
             if (p->getID() == id) {
                 if(!p->dead) {
@@ -339,7 +339,7 @@ std::vector<npc*> overmapbuffer::get_npcs_near(int x, int y, int z, int radius)
     std::vector<npc*> result;
     for(std::list<overmap>::iterator it = overmap_list.begin(); it != overmap_list.end(); ++it)
     {
-        for (int i = 0; i < it->npcs.size(); i++) {
+        for (size_t i = 0; i < it->npcs.size(); i++) {
             npc *p = it->npcs[i];
             // Global position of NPC, in submap coordiantes
             const tripoint pos = p->global_sm_location();
@@ -360,7 +360,7 @@ std::vector<npc*> overmapbuffer::get_npcs_near_omt(int x, int y, int z, int radi
     std::vector<npc*> result;
     for(std::list<overmap>::iterator it = overmap_list.begin(); it != overmap_list.end(); ++it)
     {
-        for (int i = 0; i < it->npcs.size(); i++) {
+        for (size_t i = 0; i < it->npcs.size(); i++) {
             npc *p = it->npcs[i];
             // Global position of NPC, in submap coordiantes
             tripoint pos = p->global_omt_location();
